% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bangsa_get.R
\name{bangsa_get_inventory}
\alias{bangsa_get_inventory}
\title{Get the OpenBangsamoro Database inventory}
\usage{
bangsa_get_inventory(dataset = c("geospatial", "tabular", "docs", "metadata"))
}
\arguments{
\item{dataset}{A string value from either \code{"geospatial"}, \code{"tabular"}, or
\code{"docs"} specifying the inventory to retrieve.}
}
\value{
A tibble of the specified inventory of various datasets available in
the OpenBangsamoro Database.
}
\description{
Get the OpenBangsamoro Database inventory
}
\examples{
## Retrieve the geospatial datasets inventory
bangsa_get_inventory(dataset = "geospatial")

}
\author{
Ernest Guevarra
}
