% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bangsa_get.R
\name{bangsa_get_metadata}
\alias{bangsa_get_metadata}
\title{Get the OpenBangsamoro Database metadata}
\usage{
bangsa_get_metadata()
}
\value{
A tibble containing metadata information regarding the
OpenBangsamoro Database.
}
\description{
Get the OpenBangsamoro Database metadata
}
\examples{
bangsa_get_metadata()

}
\author{
Ernest Guevarra
}
