% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bangsa_get.R
\name{bangsa_ls}
\alias{bangsa_ls}
\alias{bangsa_ls_geodata}
\alias{bangsa_ls_tabular}
\alias{bangsa_ls_docs}
\title{List files and folders within the OpenBangsamoro Google Drive database}
\usage{
bangsa_ls(id = "1W8JV4hth_gs1V6efb92HMz0eU2LRYbOO")

bangsa_ls_geodata()

bangsa_ls_tabular(id = bangsa_ls()$id[bangsa_ls()$name == "2-Tabular Data"])

bangsa_ls_docs(
  id = bangsa_ls()$id[bangsa_ls()$name == "3-Development Plans, Studies, and Reports"]
)
}
\arguments{
\item{id}{Character vector of the OpenBangsamoro \strong{Google Drive} id. This is
currently set to \emph{1W8JV4hth_gs1V6efb92HMz0eU2LRYbOO}}
}
\value{
A tibble of names and ids of files and folders within the the
OpenBangsamoro Google Drive
}
\description{
A convenience wrapper to \link{googledrive} functions set to specifically access
the OpenBangsamoro Google Drive database and list files and folders within.
}
\examples{
bangsa_ls()

}
\author{
Ernest Guevarra
}
