% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openbangsamoro.R
\docType{package}
\name{openbangsamoro}
\alias{openbangsamoro-package}
\alias{openbangsamoro}
\title{An Interface to the OpenBangsamoro Database}
\description{
The OpenBangsamoro initiative supports the use of open statistical,
geospatial, and administrative data for transparent, accountable, and
participatory decision-making as the Autonomous Region in Muslim Mindanao
(ARMM) transforms into the Bangsamoro Autonomous Region in Muslim Mindanao.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://panukatan.io/openbangsamoro/}
  \item \url{https://github.com/panukatan/openbangsamoro}
  \item Report bugs at \url{https://github.com/panukatan/openbangsamoro/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
